[Annotation] modularize
