+++ src/config/services.neon
@@ -1,3 +1,8 @@
+# Annotation
+- ApiGen\Annotation\AnnotationDecorator
+- ApiGen\Annotation\AnnotationSubscriber\ReturnAnnotationSubscriber
+- ApiGen\Annotation\AnnotationSubscriber\ParamAnnotationSubscriber
+
 # ApiGenApplication
 - ApiGen\Application\ApiGenApplication
 
@@ -39,8 +44,6 @@
 - ApiGen\Templating\Filters\Helpers\LinkBuilder
 - ApiGen\Templating\Filters\AnnotationFilters
 - ApiGen\Templating\Filters\UrlFilters
-
-- ApiGen\Annotation\AnnotationDecorator
 
 # Theme
 - ApiGen\Theme\ThemeResources
